[bridge/70] cache the result of get_mutable_bridge_object_arg (#17152)
## Description 

instead of querying the bridge arg on chain every time, we do this in
the beginning of the program and cache it. If the fetch (with retry)
fails, it panics.

## Test plan 

existing tests 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: