[sui-framework][tooling] Add human-readable published-api file (#17213)
## Description 

Generate a human-readable file containing all structs and functions that
will be published with the framework. This makes it easier to track when
functions/structs are added or removed. The output file is not meant to
capture the full signature, but simply be an easy way to determine
if/when a new function or struct is added.

## Test plan 

Built and checked the output file.