migrate indexer writer to data ingestion framework (#16680)
notes:
* a new optional method has been added to the `Worker` trait to support
the indexer workflow: `preprocess_hook`. It is blocking and enables the
execution of certain actions for a task outside of the main
`process_checkpoint` routine. It is guaranteed that `preprocess_hook`
will be called in order for all checkpoints. In this scenario, it
enables the indexing of packages before indexing the checkpoint itself