quorum-driver: add ability to call handle_certificate_v3
Refactor transaction orchestrator, quorum driver, and authority
aggregator to be able to make a call to handle_certificate_v3. This is
done by always providing a V3 request type but then selecting if we can
call handle_certificate_v2 or handle_certificate_v3 based on if input or
output objects are requested.