[docs][easy] SDKs and Move links (#17287)
## Description 

Updated the SDK page to use the card elements. We've tried to keep the
cards to just overview pages but I think because these are mostly
external links, it deserves an exclusion. Page looks weird otherwise.
Also, updated the language as it was too casual in spots and didn't go
with overall tone of site (or rest of page :). Added the Move links and
removed the toml and lock files as they are covered in move book.
Updated links to Move Book and Reference.

## Test plan 

👀 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: