[TrafficControl] Dry run mode (#17297)
## Description 

Dry run mode runs the policy without enforcing its output (does not
block requests, but will still accumulate and log, so we can see the
would-be effects on both the node and the request).

Note that this does nothing for BPF enforcement - with dry run enabled,
we may still run BPF and populate the BPF blocklist. BPF userspace will
itself need to introduce a dry-run mode in order to ignore the blocklist
and instead log/bump metrics. Until then, we will need to explicitly not
configure traffic controller to delegate to BPF (which will be a less
complete dry run).

## Test plan 

Added tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: