[GraphQL] Remove "Live" ObjectKind, rename "Historical" to "Indexed"
## Description

With the removal of direct accesses to the `objects` table, the
nomenclature we use around object kinds ("not indexed", "live",
"historical", "deleted or wrapped") is less relevant. In particular,
we never use "live" and the name "historical" is confusing because
even the latest version of the object may be considered "historical".

This PR removes the "live" kind, and renames "historical" to "indexed"
to better reflect how the users of GraphQL perceive the kinds of
objects, rather than how the data is stored in the indexer's tables.

## Test plan

```
sui$ cargo nextest run -p sui-graphql-rpc
sui$ cargo nextest run -p sui-graphql-e2e-tests --features pg_integration
```