[Consensus 2.0] extract timestamp method to a Clock component. (#17323)
## Description 

This PR refactors the method to retrieve the current system timestamp
and instead creates a dedicated Clock obj to facilitate the
functionality. The refactor will fix the failing sim tests
[here](https://github.com/MystenLabs/sui/actions/runs/8813919886)

The static initialisation of the unix epoch instant is creating issues
with simulation testing framework which is not allowing the
re-initalisation from one iteration to another, and effectively making
simtests with multiple iterations fail. Different iterations are
initialised with different dates, thus making a test run with multiple
iterations going back and forth in time effectively making Mysticeti
producing timestamps from the first run (which intialised the static
context). Failures here:
https://github.com/MystenLabs/sui/actions/runs/8813919886

With the new approach we don't deal with any static initialisation but
we rather construct from the beginning when the consensus authority node
gets created. In the future we can also easily mock the Clock object as
well for tests. Also putting the `Clock` behind the `Context` it made
the whole integration very seamless.

## Test plan 

CI/Simulation tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: