[CLI/PTB] Avoid fetching the protocol config when building a PTB
## Description

Fetching the protocol config version from RPC and then calling
`ProtocolConfig::get_for_version` can cause a problem when the network
is ahead of the CLI (it will panic). The right way to read a
particular config in this case is to read it from the RPC response
directly, but in this case, we don't really need to do it -- we were
fetching the max package size to fill a parameter when creating a
`MovePackage` but this parameter is just used to perform a bounds
check, which we don't need to do in the CLI, for a package that
already exists on-chain.

## Test plan

```
sui$ cargo nextest run -p sui
```

+ Publish a package locally and call a function against it, to make
sure package resolution is still working.