[GraphQL] Fix endpoint for GraphiQL IDE (#17483)
## Description 

Fixes the endpoint for the GraphiQL service to always be `/graphql` +
version if it is specified, regardless if the request is to `[...]/` or
to `[...]/graphql`.

## Test plan 

Visually in the browser. Checked: 
* bad version `/graphql/test` is caught: ✅ 
* good version works `/graphql` or `/` or `/stable` or `/legacy`: ✅ 
* if the routing actually goes where it needs


![image](https://github.com/MystenLabs/sui/assets/135084671/61bdb5fd-3d47-4f2b-bc1d-5874c403ec0c)

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: