[GraphQL/MovePackage] Query for latest version
## Description

Add a new kind of package point look-up to get the latest version of
the package at a given ID (or from another `MovePackage`). For system
packages, this is analogous to getting the latest version of the
object at that ID, but the versions of other packages all exist at
different IDs.

## Test plan

New transactional tests:

```
sui$ cargo nextest run -p sui-graphql-e2e-tests \
  --features pg_integration                     \
  -- packages/versioning
```