[move] Added special support for clever errors in macros (#17505)
## Description 

- Added support for `assert!(cond)` line numbers inside of macros. The
line number will point to the invocation of the macro, instead of the
line where the assert was written. This should be helpful for
cross-module macro invocations with `assert` and error codes
- Thanks @alnoki for the discussion leading to this improvement! 

## Test plan 

- Added new tests
- The line numbers are a bit off in transactional tests, so I
temporarily and localy added compiler warnings to verify the line
numbers correspond to the correct loc.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: