Add a new clippy target to run on mysql feature protected code (#17548)
## Description 

Without this PR, with xclippy we run `cargo clippy --all-features` but
that does not get run on mysql protected code since we enable only one
code path i.e. either postgres or mysql and postgres is set as default
everywhere. We want to however run clippy on mysql code paths as well.
With this, we should be able to do so.

## Test plan 

This PR