[gql] expose db_pool_size on ConnectionConfig (#17595) (#17597)
## Description 

There isn't a way to configure `db_pool_size` on `ConnectionConfig`
since it's not exposed on `start_server`. This is a simple PR that
exposes it, and bumps the default pool size from 3 to 10. This is
because we already have a few background tasks in addition to the main
graphql service.

## Test plan 

Existing tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: Expose an additional parameter `db_pool_size` on
`start-server` command so that users can configure how many db
connections the graphql service will spin up for its pool.
- [ ] CLI: 
- [ ] Rust SDK:

## Description 

Describe the changes or additions included in this PR.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: