[gql] thread cancellation token through indexer to address flaky grap… (#17733)
In the codebase today, when each e2e test is completed, we send a
cancellation token to the graphql service and then delete the database
created for the test. The issue is that the indexer and its supporting
tasks continue to run, and now that they're trying to establish a
connection to a db that does not exist, every connection will wait for
`CONNECTION_TIMEOUT` duration. This means every test hangs for much
longer than needed. To resolve, we thread a cancellation token through
the indexer's test instantiation, so that on cleanup we can send a
cancellation signal to both services, and finally spin down the db.

Manually check that test duration is shorter.

---

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:

## Description 

Describe the changes or additions included in this PR.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: