fix-windews-build-error
Due to the use of conditional compilation for the set_reuseport method, it will not be compiled on certain systems (such as Windows). Use the same conditional compilation tag to ensure that this conditional branch is also not compiled when the set_reuseport method is not compiled. This can fix compilation errors on specific systems.