[cherry-pick][graphql] Cherry pick windows build fix to 2024.4 (#17800)
## Description

A file was accidentally checked in that isn't compatible with Windows --
our Windows build CI is failing, and we also have reports from users
that it's messing with their git checkouts on Windows. I don't think the
file is supposed to be there, so let's remove it.

Fixes #17375 

## Test plan

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

Co-authored-by: Ashok Menon <ashok@mystenlabs.com>