[13/n][enums/Sui] Sui e2e tests for Move enums (#17251)
## Description 

This is the top-level (and final!) PR in the stack for Move enums.

This PR
* Updates the protocol config to support the new execution version
* Adds new e2e adapter tests to make sure we handle enums properly
* Updates snapshots for tests (and protocol config snapshots and the
like).
This also then tests all other tests that have been added below this in
the stack (e.g., the package resolver tests, graphql tests, all the Move
tests, other misc tests that were added).

**NB:** Before landing this we need to determine the best way of turning
on the correct binary-version for testing only. So I would encourage
mainly looking at the tests (and the updated snapshots for the replay
tool). Protocol-config changes, and snapshots for them will need to be
updated as this gets closer to landing, and we further concretize the
testing/pre-rollout/rollout plan.

## Stack:
* #17245 
* #17246 
* #17247 
* #17248 
* #17249 
* #17250 
* #17251 **<< You are here** 🎉 

## Test plan 

It's pretty much all tests.