[8/n][enums/Sui] Move enums in Sui execution (#17248)
## Description 

Changes for Move enums in `sui-execution`. Generally nothing surprising.
Only thing worth note is that enums are not allowed to be OTWs (see
tests in top commit in this stack).

This PR is meant to be a logically-reviewable portion but is not
land-able on its own. It must be merged in with the changes in the rest
of this stack to be landed.


## Stack:
* #17245
* #17246 
* #17247 
* #17248 **<< You are here**
* #17249 
* #17250 
* #17251 

## Test plan 

Tests in the top commit on this stack.