Add VDF to sui framework (#16196)
## Description 

This adds a VDF verifier (Wesolowski's construction based on imaginary
class groups) to the Sui framework. It can be used to generate on-chain
randomness in an 1-out-of-n trust model (see for example [this
paper](https://crypto.stanford.edu/~dabo/pubs/papers/VDFsurvey.pdf) for
an explanation).

This PR also includes an example of a lottery using VDF to generate it's
randomness.

Boilerplate code and snapshots will be added after the review.

Closing #7222.

## Test Plan 

Unit tests.

### Type of Change (Check all that apply)

- [X] protocol change
- [X] user-visible impact
- [X] breaking change for a client SDKs
- [X] breaking change for FNs (FN binary must upgrade)
- [X] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Add verifiable delay functions (VDFs) to the Sui framework. The VDF is
based on Wesolowski's construction over imaginary class groups. Note
that this is supported in Devnet only.

---------

Co-authored-by: Todd Nowacki <tmn@mystenlabs.com>
Co-authored-by: Ashok Menon <ashok@mystenlabs.com>