[move][naming][0/k] Clean up naming infrastructure (#17839)
## Description 

This cleans up and combines a lot of the code in naming into a
more-uniform interface. This work is in service of hopefully one day
combining path expansion and name resolution into a single operation.

## Test plan 

A few of the tests were updated with expected output. Bikeshedding is
quite welcome on those, but overall things are more-uniform with more
normalized structure now.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: