[indexer] Move `exchange_rates` to be a free function (#17983)
Move exchange_rates to be a free function, and remove unneeded APIs.
GraphQL APY code is reworked in #17701.
 
This PR needs to be merged first, before #17701. Note that the diff
messes things up. The code was not changed, except for `self -> state`.

## Description 

Describe the changes or additions included in this PR.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: