Run framework verification on its own VM (#18027)
## Description 

Gated by a protocol version we run upgrades of framework in its own VM.
That is to avoid loader cache issues when verification of a VM can use
packages that are loaded already and as such incompatible with the new
upgrade.
This solution should solve the vast majority of problems related to
changes in framework code.
Next steps:
- forbid new `init` in framework modules (verifier check)
- make a easy to use framework to push `init` style code on the epoch
after the framework upgrade. That is in line with what has been done for
all new systems we exposed in the framework (e.g. random beacon and
bridge)

## Test plan 

Existing tests and about to run multi-epochs replay

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: