[Consensus 2.0] Fix retries and wait time during fetch block request timeout (#18033)
## Description 

Retries were not increased when a request was timing out which probably
causing issues when trying to sync with nodes that have their ports
blocked. This PR is increasing the retries + adding some wait time for
it to ensure that it's graceful.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: