[TrafficController] Fix fw delegation tests (#18106)
## Description 

There are two classes of errors from what I've seen: 
1. `Expected blocklist to be non-empty` -- this was simply due to race
conditions between when the test server garbage collects IPs from the
blocklist after TTL and when the test checks for existence in the
blocklist. If there's some sort of network delay then we may check for
existence too soon. To fix, we could either increase the TTL timeout or
have signaling between the server and the test (note that we can't run
as a simtest today because simtests don't correctly simulate socket
connections). I chose to increase the TTL and remove the later garbage
collection check as it's not really necessary (it's really only checking
that we correctly pass the TTL config to the "remote server"). However
if this continues to be an issue, can add signaling/message passing.

2. `error binding to 127.0.0.1:65000: error creating server listener:
Address already in use (os error 98)` -- this should be as simple as
running each server configured to bind to a different port, in case they
happened to be spun up concurrently.

## Test plan 

Run the tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: