[chore][GraphQL] Declutter schemas
## Description

Remove `draft_target_schema.graphql` and promote
`current_progress_schema.graphql` to be the canonical schema for the
service -- move it to the top-level of the `sui-graphql-rpc` crate to
make it easier to find. This is to avoid confusion about source of
truth for the GraphQL schema.

Because the TS SDK references the schema at multiple GraphQL versions,
we will need to cherry-pick this change to release branches when it
lands.

## Test plan

CI