[move][move-2024][enums] Fix a small issue when match is malformed (#18387)
## Description 

This fixes a bug that was previously a compiler crash when a match was
malformed and counterexample generation occurred. This was due to some
changes for parser resilience combined with moving match analysis to
typing.

## Test plan 

Added the repro test, and it now produces the error as expected.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: