Skip submitting checkpoint signatures when checkpoint is already certified (#18398)
## Description

During catchup in mysticeti if the validator was down for long periods
of time we will see a large number of checkpoint signatures start to get
[backlogged](https://metrics.sui.io/goto/HgmncvwSR?orgId=1) and then
submitted to consensus quickly after the node has caught up. This will
cause the block size to increase in consensus and then a variance in the
distribution of those blocks across the network leading to a round rate
dip. This PR will skip submitting checkpoint signatures when checkpoint
is already certified to mitigate the issue and we will look into
properly handling block size variance in a separate PR

## Testing

Recreated the issue and tested the fix in private-testnet

![Screenshot 2024-06-25 at 5 15
33 PM](https://github.com/MystenLabs/sui/assets/97870774/b3be4e36-6e56-4c28-a231-f2a43a358118)