[bridge]  support rank validators by voting power when requesting signatures (#18298)
## Description 

In this PR we introduce a way to order validators to request signatures
according to their voting power. This is useful for signatures to be
verified on ethereum because the # of signatures do impact the gas cost
there significantly.


## Test plan 

deploying on a testnet bridge node to test.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: