[bridge] use protocol parameter to determine when to finalize bridge committee `bridge_should_try_to_finalize_committee`  (#18019)
## Description 

Updated: we decided to introduce a protocol parameter to determine when
bridge committee happens. This gives us more flexibility to include
validators. As a result, we don't need the original change of threshold
any more. Together with the new parameter, the old default value 7500
still acts as the minimal voting power threshold


------------------

Old description:


Previously the 75% minimal stake is too low to include most validators.
This PR changes it to 90%.

I have two commits in this PR:
* the first commit sets the value by distinguish chain id - if it's
testnet, then use 7500, otherwise 9000. This is safe because on mainnet
we haven't enabled registration yet.
* the second commit uses a different approach with protocol config.
Basically it adds a `minimal_threshold_for_bridge_committee` field and
is set to 90% after the added version. In this way we don't need to
differentiate chain ids. It's safe for testnet because this value won't
be needed post committee initialization.

I like the 2nd commit better because the code is cleaner.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: