[RFC][core] Add validator tx finalizer (#18542)
## Description 

The ValidatorTxFinalizer gets called whenever a transaction is signed on
a validator. It would then sleep for a min, wake up and check if the tx
has already been executed, if not, use authority aggregator to finalize
it.

## Test plan 

Added unit tests to cover various cases:
1. Basic flow
2. Epoch change
3. Tx already executed
---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: