[pick #18463, #18573][Examples/Move] Fix Move Example test set-up
## Description

All Move examples should depend on the Sui framework in the repo, so
that this framework is compatible with the compiler that is being used
to compile them -- especially when running CI on non-main branches.

## Test plan

```
sui-framework-tests$ cargo nextest run
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK: