Integrate validator tx finalizer (#18681)
## Description 

This PR integrates the validator tx finalizer with the validator
service, which is then invoked after handling a transaction.
A node config is used to control whether an instance of the service is
created when the service starts.
It is enabled on devnet and testnet. We will enable for mainnet later
when we feel confident.

## Test plan 

Added e2e simtests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: