[bridge] add metrics for SignerWithCache (#18685)
## Description 

as title. `SignerWithCache` today already guarantees single processing
of one request (if they are received in a short period of time and not
evicted in cache). In this PR we add hit and miss metrics for us to
understand how well it works.


https://github.com/MystenLabs/sui/blob/main/crates/sui-bridge/src/server/handler.rs#L152


## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: