[Consensus 2.0] Recover from amnesia - part 2 (#18190)
## Description 

The second part of the recovering from amnesia feature. This PR is
implementing the logic to fetch the last own block from the peer network
in the synchronizer. It is also adding several tests to confirm the
behaviour correctness.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: