[docs] Content updates and style changes (#18745)
## Description 

Addresses requests made by aslan. Adds a style guide entry for using
code in headings. Previous opinions were 50/50 and he tipped the scale.

Original request was to remove the bullets in
`swap_exact_quote_for_base` to make it the same as the previous
description. But the suggestion is to keep the bullets to break up the
content and make the object list easier to read.

Also breaks up the last paragraph on Design because it looked weird. 

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: