[bridge] add BridgeMonitor and handle url change event (#18790)
## Description 

This PR adds `BridgeMonitor` which receives all `SuiBridgeEvent` (and
probably `EthBridgeEvent` as well soon) and handles them accordingly. In
this PR we add the handling for `CommitteeMemberUrlUpdateEvent` by
retrieving the latest committee onchain and swap it in.

## Test plan 

added unit tests


---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: