[gql][indexer] index chain identifier into its own table (#18825)
This PR puts chain identifier into its own table so that queries of
chain identifier does not depend on the existence of checkpoint 0 in the
db, which may be pruned away.

Tested against devnet locally and added a gql e2e test.

---

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [x] GraphQL: Added a way to always have `chainIdentifier` query return
the correct chain ID even when pruning is enabled.
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API:

---------

Co-authored-by: stefan-mysten <135084671+stefan-mysten@users.noreply.github.com>
Co-authored-by: Ashok Menon <ashok@mystenlabs.com>