[Indexer] for_all_tables skips explicit partitions
## Description

Omit the partition tables from the `for_all_tables` list, because we
don't want to refer to them directly (outside of the partitioned table
they are part of).

This fixes an issue with GraphQL's compatibility check which would fail
because it tried to read from the partition table directly, and it isn't
there on a DB served by an indexer with pruning enabled.

## Test plan

Start GraphQL service against a pruned DB.