rocksdb: access rocksdb through typed-store
This patch does some dependency shuffling to force accessing the rocksdb
crate via the typed-store crate in order to try and limit how and when
librocksdb-sys is built.

Before this patch, when building all targets, librocksdb-sys was
included in the dependency graph and built twice (with a total of 2005
targets). After this patch librocksdb-sys is included only a single time
in the dependency graph and reduces the number of built targets to 1959.