rocksdb: reintroduce workspace-hack scoped to typed-store
Reintroduce a workspace-hack package (`typed-store-workspace-hack`) in
order to perform feature unification specifically for the third-party
dependencies which impact the building of rocksdb, the heaviest weight
dependency we currently have.

After this patch, no matter where in the workspace you perform a build,
rocksdb will only be built a single time.

For now the `typed-store-workspace-hack` package is manually maintained
and was created with assistance from `cargo hakari` and `cargo tree`.