[zkLogin] Check that bigints fit into the field for extra safety (#16891)
## Description 

Turns out we don't check anywhere that the inputs fit into the field and
neither does the poseidon-lite library (it just truncates elements and
fits them into the field).

Instrumenting poseidon seems like the simplest way to do this. The other
option (which I have not taken) is to instrument as soon as we receive
some inputs..

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes