fix(verify-source): Detect published-at = 0x0
## Description

Detect when the `published-at` field in `Move.toml` or `Move.lock` has
been explicitly set to `0x0` and treat that as if it was not set at all.

This is not commonly done by people, but it happens in our test set-up.

This also required converting the field into an `ObjectID` earlier in
the pipeline, which introduced some further changes in the codebase.

## Test plan

```
sui-source-validation$ cargo nextest run
```