[LocalExecution] Remove local execution loop from FN (#19032)
## Description 

We are about to deprecate local execution. This PR gets rid of the loop
that subscribes effects from quorum driver and execute them locally.
This was an optimization and it should be safe to remove at this point.
Removing this also allows us to see clear metrics on the number of local
execution requests received from the network.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: