fix(jsonrpc): Flaky fullnode JSON-RPC test
## Description

`transaction_tests::test_get_fullnode_transaction` included a flaky
assertion that if you fetch 5 transactions, then the rest of the
transactions, and then fetched the latest 10 transactions, the latter
would match up with the suffix of the former two queries concatenated
together.

This does not work in general (at least not with this test set-up),
because of consensus commit transactions.

I have updated the test to check a similar property: That if you fetch 5
transactions, then the rest, and then fetch the first 10 transactions,
then the prefixes match (this tests that transaction order is stable).

There are other existing tests for what happens when you query
transactions in descending order.

## Test plan

This seed previously exhibited the failure, and now it succeeds.

```
sui-json-rpc-tests$ MSIM_TEST_SEED=968774516445189525 cargo simtest -- test_get_fullnode_transaction
```