unit test (#18659)
## Description 

Unit test infra for the bridge.
Not sure why `bridge-env.move` is not showing as a rename.
Also I run the latest formatter on the test files and that may add to
the change.
In any case @patrickkuo @Bridgerz we will talk about this given that is
the easiest way to understand what is going on.

Overall it is a purely additive change.

## Test plan 
This is it. Old tests run and new test increase coverage (~90%). We need
to write few tests to cover for some assertion and that is coming

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: