fixup: use half-open intervals everywhere.
There was a bug related to bounds for the first page, when it contains
the genesis checkpoint. This was fixed by refactoring `TxBounds` to
systematically use a half-open interval (inclusive lowerbounds and
exclusive upperbounds) when calculating and combining intervals to be
used on the DB, regardless of where the bounds came from and whether
they were inclusive or exclusive at source.

Also added a new test to exercise this behaviour -- all existing tests
also pass, unchanged.