[walrus/sui-proxy] example of extending sui-proxy for additional use cases
## Description
this pr demos how to add additional routes to the server without changing or affecting the existing behvior of sui node.  It allows for limitless additions and configuration in distinct code paths isolated from existing behaivor and will reduce the likelihood of breaking other parts of the proxy when extending the features.

## Test Plan

cargo fmt