[bridge-indexer] fix duplicated task creation (#19171)
## Description 

In function `tasks()` we return only incomplete tasks. As a result, when
the latest backfill tasks are done and ignored, we mistakenly use an
intermediate value as the `latest target checkpoint` and use it to fill
gaps. This causes duplicated tasks. This PR fixes it:
1. rename `tasks` to `get_ongoing_tasks` for semantics
2. add `get_largest_backfill_task_target_checkpoint` and use that to
determine whether there is a gap
3. simplify `update_tasks` for backfill task creation
4. add some utils functions for testing.



## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: