[trace-view] Added support for stepping back through the trace (#19215)
## Description 

This PR adds the ability to step back through the trace. The main
limitation is that stepping back into or over (previously executed)
function calls is not supported. We will add this support after support
for viewing variables is added as it will requires snapshotting variable
state (which we do not have at the moment).

## Test plan 

Tested manually that:
- viewer correctly steps back within the function and from inside the
callee to the caller
- viewer correctly stops stepping back at the beginning of the trace
- viewer correctly stops stepping back upon encountering previously
executed function call