Push metrics bridge (#19077)
## Description 

This adds the ability for a bridge node operator to push metrics to an
external metrics proxy for aggregation. The implementation is mostly
copied from
https://github.com/MystenLabs/sui/blob/main/crates/sui-node/src/metrics.rs

## Test plan 

Tested in testnet

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: