indexer: script to regenerate schema
## Description

Create a script that regenerates the schema and run it to standardise
our schema. Also merge `schema/mod.rs` and `schema/pg.rs` into one
`schema.rs` -- otherwise we have an opportunity for things to go out of
sync because a module is present in the generated schema (`pg.rs`) but
not in the module entrypoint (`mod.rs`).

## Test plan

```
sui$ ./scripts/generate_indexer_schema.sh
```