indexer: ensure schema always matches what diesel generated (#19337)
## Description 

Create a script that regenerates the schema and run it to standardise
our schema. Also merge `schema/mod.rs` and `schema/pg.rs` into one
`schema.rs` -- otherwise we have an opportunity for things to go out of
sync because a module is present in the generated schema (`pg.rs`) but
not in the module entrypoint (`mod.rs`).

Run this script to refresh the schema (some inconsistencies already
crept in!) and set it up to run under CI.

## Test plan 

```
sui$ ./scripts/generate_indexer_schema.rs`
```

and CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: