[move] Revise dead code detection (#19332)
## Description 

Dead code was written without respecting named blocks and loops, leading
to false negatives in cases that should work just fine. This revises.
that analysis to do the correct thing, though the implementation is
still a bit rough and full of heuristics. The main idea is that instead
of tracking "diverged" versus "did not", we track "might have" versus
"did" and "did not." This allows us enough breathing room to avoid
saying we diverged when that divergence was guarded by conditional
control flow.

In writing tests, I also found an issue in how HLIR lowered named
blocks, due to conflicting assumptions about how things are typed. This
is fixed now, allowing this sort of code:

```
fun test(): u64 {
 'a: {
     loop { return 'a 5 }
   }
}
```

## Test plan 

A bunch of new tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: