[move][vm_rewrite] package loader (#19153)
## Description 

This is the base rewrite for the package loading and caching mechanism.
Before progressing much further than this, we will need to rip out the
current loading and module/type/function resolution in the interpreter
and move over to passing vtables.

## Test plan 

This is not tested across the whole test suite. However I have added a
number of unit tests for the loader in `package_cache_tests.rs`.