[docs] Beta tag for graphql (#19358)
## Description 

Adds a beta tag to the frontmatter. Applied the tag to the graphql
topics. If networks provided as value, theyre added to the admonition
box. Format is
```
---
beta: <list of networks available in or `true`>
---
```

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: