fix(core): deserialize malformed types in tx
## Description

Make it possible to represent a transaction that includes a bad type (a
type that does not parse correctly). Such transactions will fail to
execute but may exist on-chain and so we need to be able to represent
them.

To avoid confusion in future, we will perform an early (signing) check
for type correctness, preventing more examples of this pattern from
appearing on-chain.

GraphQL has also been updated so that it can represent such a
transaction without emitting an error.

## Test plan

```
sui$ cargo simtest
sui$ cargo nextest run -p sui-graphql-rpc
sui$ cargo nextest run -p sui-graphql-e2e-tests
```