[Indexer] Add json serialized SuiSystemStateSummary to epochs table (#19428)
## Description 

The existing system_state column is a BCS serialization of a JSON-RPC
type, which is not evolvable.
This PR adds a new column that is the JSON serialization of
SuiSystemStateSummary without BCS.

It also fixes two bugs along the way:
1. We were storing the wrong version of the system state, offsetting by
1 epoch. This PR fixes it by storing the right version of the json
column at the beginning of the epoch instead of at the end.
2. We were ignoring a deserialization error and swallow it in some
place. Made that throw instead.

## Test plan 

Added a debug assert to see that this column is populated with the
correct epoch.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: