indexer: add indices to aid pruning
## Description

All the transaction and events lookup tables are missing indices that
allow the pruner to efficiently pick out the ranges of rows to remove.

Note that because these indices are being added to existing tables, they
are being added using `CONCURRENTLY IF NOT EXISTS`, and because of that,
they need to each be in their own migration file (concurrent index
creation cannot go in a DB transaction).

## Test plan

```
sui$ ./scripts/generate_indexer_schema.sh
```