chore(graphql): deduplicate configs and clap args
## Description

Avoid duplicating fields for configs that are accepted as flags from the
command-line and can also be read from TOML file. Use the same struct
for both purposes but deocrate it with both `clap` and `serde` (or
`GraphQLConfig`) annotations so that it can serve both purposes.

## Test plan

This change should preserve the config/command-line interface -- checked
by running the GraphQL service with the same invocation, referencing all
flags, before and after the change.