core: event sending module relocation
## Description

Make it so that the module that is associated as the "sending module"
for an event is relocated by linkage. This fixes an issue (captured in a
regression test) where if the function that is called by the PTB is from
some upgraded version of the package, it cannot be found because the
Event's package ID still points to the original version of the package.

## Test plan

Introduced a regression test -- before the change, the "sending module"
in the response returned `null`.

```
sui$ cargo nextest run -p sui-graphql-e2e-tests -- sending_module
```