fix(backfill): chunk up writes to DB
## Description

When a given range of transactions generated too many affected object
IDs, the backfill script would stop, because of limits in postgres'
protocol over the wire.

By chunking up the writes into batches of 1000, we avoid hitting this
limit.

## Test plan

Running this backfill against the DBs now -- and it is not crashing.