[data ingestion] introduce reducer functionality (#19656)
## Description 

The PR introduces a new `reduce` functionality to the data ingestion
framework.
The functionality is optional and will not impact existing workflows. 
When enabled, it allows for the accumulation of a batch of checkpoints
for post-processing.
The watermark in the progress store is now updated after the reducer has
completed its job.
The old method, `Worker::save_progress`, is deprecated in favor of this
approach.

As a proof of concept and to demonstrate the functionality, the archival
workflow has been migrated to utilize a reducer. Note that the previous
implementation would still work, but using a reducer for batch
aggregation is more intuitive compared to accumulating state within a
worker

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: