types: introduce FullCheckpointData enum
Introduce the FullCheckpointData enum, and CheckpointDataV2, types to:
    - enable type evolution into the FullCheckpointData format. This
      will allow us to add new information, like auxiliary data, if/when
      it is output from execution.
    - Dedup user signatures which are included twice in the
      CheckpointData format, once in the checkpoint contents and once in
      the transaction itself.

This also updates the Rest endpoints to return this new
FullCheckpointData enum instead of the CheckpointData type. In order to
limit breakage today, clients consume the FullCheckpointData enum and
simply convert it into the old CheckpointData format. In the future we
can migrate to using the new format everywhere.