donotland(indexer): increase rows written.
## Description

Experiment to increase the number of rows written across a variety of
tables. The idea is to selectively enable some of these double writes
and measure whether this impacts indexer performance.

Double writes are all marked with a sentinal value (a leading `0x42`)
which should make it easy to identify these values and clean them up
after the experiment.

## Test plan

Tested locally.