donotland(indexer): bloat transaction payload
## Description

Intentionally bloat the size of transaction records to increase bytes
sent over the wire. Random bytes are used to ensure the bytes don't get
compressed down to nothing.

This experiment is harder to revert after the fact, but it can be done
by recalling the range of transactions that the experiment ran for and
removing the last element of `object_changes` as long as it starts witht
`\xA11CEBOB...`.

## Test plan

Run locally.