indexer-alt: ingestion service
## Description

A service the fetches checkpoints and pushes them down to subscribers to
process.

## Test plan

Created a dummy ingester that prints the checkpoint sequence number and
digest to test throughput. This was run for 100s on my local machine at
two different points in mainnet's history, around checkpoint 9M, and
checkpoint 64M. The former is around the time of Sui8192 and when
on-chain checkpoint rate was around 1 per second, and the latter is from
the FOMO mint when checkpoint rate on-chain was around 5 per second:

```
sui$ cargo run -p sui-indexer-alt --release --          \
  --remote-store-url https://checkpoints.mainnet.sui.io \
  --start-checkpoint $CP
```

Locally, this simple ingester was limited in both cases by bandwidth
(around 100Mbps), which translated to a checkpoint rates of roughly 70
and 160 per second at checkpoints 9M and 64M respectively.

In a follow-up PR, I introduce an equivalent dummy ingester to the
existing data ingestion framework, and it reached rates of around 50 and
130 per second. It's possible that this is still due to bandwidth,
because the existing framework can end up dropping a partially
downloaded checkpoint if it times out (and has quite an aggressive
timeout).