[move][tracing] Cleanup insertion points of tracer in the VM (#19452)
## Description 

This cleans up the insertion points of the tracing in the VM. This does
this by moving the calls out into macros, and gating them on a const
that is redefined based on the feature flags enabled.

Note that this still keeps the gas profiler profiler calls within these
macros (but as far as the interpreter is concerned, they're gone), as
the unification of the gas profiler with the trace format can be done
later.

NB: the `tracing.rs` file has not be touched aside from removing the
feature flag and a couple clippy fixes. So even though it shows up as a
big change to the file it is in fact very small..

This also removes the (now unused) `gas-profiler` feature flag from the
`move-vm-types` crate.

## Test plan 

Make sure existing tests pass. 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: