indexer-alt: committer watermarks
## Description

Introduce the watermarks table, and update the committer to update it,
taking into account out-of-order writes to the underlying tables.

This change also allows the indexer to pick up where it left off by
consulting the watermarks table for the high watermarks it had
previously written.

It also introduces the ability to configure the range of checkpoints
being indexed (set an upper and lowerbound).

The metrics introduced by this change can be used to track per-pipeline
checkpoint rate, epoch rate and transaction rate.

Some TODOs have also been added for various tests and experiments that
should be run based on this work.

## Test plan

A lot of manual testing. Some TODOs have been left for more extensive
testing once there is more of a test set-up (depends on tempdb and
auto-migration work).