indexer-alt: simplify event indexing
## Description

Instead of having a separate table for each component of the cascading
index, have a single table, and add multiple indices to it for each of
the cascading cases.

This should reduce the footprint and ingress to the DB, but mildly
increases the risk that the DB picks a bad query plan.

## Test plan

Run all the existing tests, and also run an experiment to confirm that
the DB can successfully plan queries against this kind of schema.

Our initial fear was that if we had multiple indices on a single table,
then the DB may pick the wrong index, and there is still a chance that
might happen if we add indices for disparate filters to the same table
(i.e. we combine the event emit module and event struct instantiation
fields into one table), but we can guarantee that the reader will only
issue one query to each of these merged tables, and it should entirely
overlap with one of its indices.