indexer-alt: introduce regulator for ingestion
## Description

The regulator is responsible for stopping the flow of checkpoints if it
runs too far ahead of a high watermark. The watermark is optionally set
by any ingestion subscriber, and will be used by pipelines that need to
commit data in checkpoint order (like the `objects`/`objects_snapshot`
pipelines) to prevent their internal buffers from growing without bound.

This complements the existing back-pressure that's based on channel
buffers, which is a useful mechanism for out-of-order/concurrent
pipelines, but not so useful for in-order pipelines.

## Test plan

The regulator is not used by any existing pipelines, but has been
plumbed into the indexer and new tests have been added:

```
sui$ cargo nextest run -- ingestion::regulator
```