indexer-alt: pull out watermark task (#19943)
## Description

Move logic for updating watermarks out from the committer and into its
own task. This serves multiple purposes:

- Simplifies the committer
- Prepares the committer for supporting multiple concurrent writes (each
write can then push its watermarks to this task to accumulate and
update).
- We may be able to re-use some of these pieces in a sequential pipeline
(this is not a sure thing because the sequential pipeline may have
slightly different requirements for all parts).

## Test plan

Ran the indexer with the following configurations:

- With no limit and shutdown by interrupt,
- From scratch, up to checkpoint 3000,
- Between checkpoints 9,000,000 and 9,002,000,
- With and without `--skip-watermarks` enabled,
- From a clean DB and on a re-run.

And ensured it made progress and successfully shutdown at the expected
time in each case.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: