indexer-alt: concurrent writes
## Description

Separate out the concurrent committer one more time: Into a collector
and a committer, and add support for concurrent writes to the new
committer.

To support this we need to also deal with the fact that the watermark
task might receive the watermark for a checkpoint that is only partially
written out, which is done through the introduction of the
`WatermarkPart` -- this tracks a watermark as well as a fraction of the
rows that come from the checkpoint at the tip of this watermark.

## Test plan

Run the indexer pipeline from scratch up to checkpoint 5000, re-run it,
run just a single pipeline, and run it at some later checkpoint as well
(e.g. 9,000,000 to 9,002,000) and make sure it exits cleanly after write
a complete data stream, in all cases.