indexer-alt: fix connection timeout
## Description

Using `stringify!(DEFAULT...)` doesn't work, because it stringifies the
constant name, not its contents. This change uses `const_str::format` to
do the same thing instead.

## Test plan

Run the indexer without supplying a connection timeout. Before this
change it would fail with a parse int error, because it tried to
interpret the string `DEFAULT_CONNECTION_TIMEOUT_SECS` as an integer.