ts sdk: coin-with-balance intent create zero coin on zero amount (#20184)
## Description 

Change the `coinWithBalance` intent so that it creates a zero coin
instead of splitting when amount is 0. This fixes the edge case where
the user doesn't have a coin of required type in the wallet but the
amount is 0 anyways, thus avoids throwing the `Not enough coins of type
${coinType} to satisfy requested balance` error in this situation

cc @hayes-mysten 

## Test plan 

Added e2e tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API:

---------

Co-authored-by: hayes-mysten <135670682+hayes-mysten@users.noreply.github.com>