indexer-alt: watermarks.pruner_timestamp is a TIMESTAMP
## Description

Change `watermark.pruner_timestamp_ms BIGINT` to
`watermark.pruner_timestamp TIMESTAMP`. This is because unlike the other
timestamps in this table, which are read from checkpoint data, this
timestamp is read from the database, and so it is easiest to represent
it using the database's native type for timestamps.

## Test plan

CI