indexer-alt: reader watermark task
## Description

The reader watermark task is one half of the pruning implementation. It
applies only to concurrent pipelines, and tracks the lowest checkpoint
sequence number the reader is guaranteed to get data for, and the
timestamp (from the DB), that this threshold was set at.

This task does not do any actual pruning, but it sets the limit that the
pruner can go to, and a time at which it can start pruning to that
point. A later PR will introduce the actual pruning task.

Unlike the existing pruner implementation, this system does not support
pruning by epoch, only by checkpoint. This was for two reasons:

- Pruning no longer works by dropping partitions, so we can keep the
  implementation simple by only supporting the form of pruning we need
  (by checkpoint sequence number).
- Pruning is part of the indexing framework, rather than a specific part
  of our indexer implementation, so pruning by epoch would require that
  the indexing framework kept track of a mapping from epoch to
  checkpoint sequence number.

Along with the reader watermark, the following changes were made:

- The existing watermark task was renamed to the `committer_watermark`
  to distinguish it.
- The interface between pipelines and the indexing framework has been
  simplified to return just one `JoinHandle`. This was because clippy
  complained that returning 6 handles was too complex, and additionally,
  because the reader watermark task is not connected to the rest of the
  tasks for a pipeline, it needs a more complicated tear down logic,
  where firstly the committer tasks are wound down, and then the pruner
  is explicitly cancelled, before its tasks are awaited as well.

The change also sets up pruning configs for write-ahead log tables, so
that we can test reader watermark behaviour.

## Test plan

Run the indexer with just one of the consistent read summary and
write-ahead logs, and then query the watermarks table:

```
cargo run -p sui-indexer-alt --release --                                        \
  --database-url "postgres://postgres:postgrespw@localhost:5432/sui_indexer_alt" \
  indexer --remote-store-url https://checkpoints.mainnet.sui.io                  \
  --last-checkpoint 10000                                                        \
  --consistent-range 100 --consistent-pruning-interval 10                        \
  --pipeline sum_obj_types --pipeline wal_obj_types
```

```
sui_indexer_alt=# SELECT * FROM watermarks;
   pipeline    | epoch_hi_inclusive | checkpoint_hi_inclusive | tx_hi | timestamp_ms_hi_inclusive | epoch_lo | reader_lo |      pruner_timestamp      | pruner_hi
---------------+--------------------+-------------------------+-------+---------------------------+----------+-----------+----------------------------+-----------
 sum_obj_types |                  1 |                    9900 |  9902 |             1681405380769 |        0 |         0 | 1970-01-01 00:00:00        |         0
 wal_obj_types |                  1 |                   10000 | 10002 |             1681405504878 |        0 |      6321 | 2024-11-10 17:33:11.459593 |         0
(2 rows)
```

Note that `reader_lo` is not guaranteed to be as advanced as it could
be, because this task checks in relatively infrequently, and the indexer
as a whole might shut down efore it has a chance to catch up. However,
on restart, the indexer should fast-forward the watermark (i.e. if the
steps above are repeated, `reader_lo` will be updated):

```
sui_indexer_alt=# SELECT * FROM watermarks;
   pipeline    | epoch_hi_inclusive | checkpoint_hi_inclusive | tx_hi | timestamp_ms_hi_inclusive | epoch_lo | reader_lo |      pruner_timestamp      | pruner_hi
---------------+--------------------+-------------------------+-------+---------------------------+----------+-----------+----------------------------+-----------
 sum_obj_types |                  1 |                    9900 |  9902 |             1681405380769 |        0 |         0 | 1970-01-01 00:00:00        |         0
 wal_obj_types |                  1 |                   10000 | 10002 |             1681405504878 |        0 |      9801 | 2024-11-10 17:33:49.913129 |         0
(2 rows)
```