chore(deny): remove direct dep on derivative
## Description

`derivative` is unmaintained. Switch existing uses to use `derive_more`
instead, which is an existing dependency that has all the necessary
features.

This also required bumping `derive_more` to `1.0.0`, which introduced
some small breaking changes, and means that we are pulling in two
versions of `derive_more` due to transitive dependency constraints, but
in this case, it should be okay because we are mainly using this dep for
its proc macros, and one hopes that eventually the transitive deps will
upgrade to `1.x` as well.

## Test plan

CI