[Consensus] resubmit GC'ed transactions (#19572)
## Description 

Transactions of GC'ed blocks should be retried to remain compliant with
the so far system assumptions. Refactored the consensus side to signal
back the submitter about the status of the block that the submitted
transaction has been included to. Practically two possible outcomes:
1. Sequenced (when the block has been committed and consequently the tx
as well)
2. GarbageCollected (when the tx's block has not been committed and
passed gc_round)

also took the opportunity for a small refactoring around the
`SubmitToConsensus` trait and the `ConsensusClient` as intentions
started diverging.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: